feat(i18n): 重新组织翻译提供者组件，提升代码结构和可维护性
